Fix merged modules build
* Fix some merge conflicts that weren't resolved quite right
* Fix imports with the module syntax
