+++ cravat/websubmit/nocache/core.js
@@ -191,6 +191,10 @@ export {
     addEl,
     changePage,
     PubSub,
-    OC
+    OC,
+    titleCase,
+    getSizeText,
+    emptyElement,
+    removeElementFromArrayByValue
 };
