+++ cravat/websubmit/nocache/main.js
@@ -1,4 +1,5 @@
 import * as core from './core.js';
+import * as moduleInfo from './moduleinfo.js'
 import * as submit from './websubmit.js';
 import * as store from '../../store/nocache/webstore.js';
 import * as header from './header.js';
