+++ cravat/websubmit/nocache/websubmit.js
@@ -1,10 +1,14 @@
 'use strict';
 import {
-getTn, addEl, getEl, changePage, PubSub, OC
+    getTn, addEl, getEl, changePage, PubSub, OC, titleCase, emptyElement
 } from './core.js'
 import {
     connectWebSocket, checkConnection, getBaseModuleNames, toggleChatBox
 } from '../../store/nocache/webstore.js';
+import {
+    getRemote, populateAnnotators
+} from './moduleinfo.js';
+import { loadSystemConf } from './header.js'
 
 // var OC = OC || {};
 OC.servermode = false;
